+++ Source/Universum/Universum/World/CelestialObject.cs
@@ -31,7 +31,6 @@ namespace Universum.World {
         protected Quaternion _rotation = Quaternion.identity;
         protected bool _addBillboardRotation = false;
         protected Quaternion _billboardRotation = Quaternion.identity;
-        protected Quaternion _billboardCorrectionRotation = Quaternion.identity;
         protected Quaternion _axialRotation = Quaternion.identity;
         protected Quaternion _spinRotation = Quaternion.identity;
         protected Quaternion _inclinatioRotation = Quaternion.identity;
@@ -46,16 +45,16 @@ namespace Universum.World {
         public Vector3 scale;
         protected float _scalePercentage;
         public float extraScale;
-        public float speed;
-        protected float _speedPercentage;
+        public double speed;
+        protected double _speedPercentage;
         protected int _period;
         protected int _timeOffset;
         protected float _orbitPathOffsetPercentage;
         protected float _orbitEccentricity;
-        protected float _orbitSpread;
+        protected double _orbitSpread;
         protected int _orbitDirection;
-        protected float _spinRotationSpeed;
-        protected float _orbitRadius;
+        protected double _spinRotationSpeed;
+        protected double _orbitRadius;
         protected float _yOffset;
 
         public CelestialObject(string celestialObjectDefName) {
@@ -68,7 +67,7 @@ namespace Universum.World {
                 objectHolder.Destroy();
             }
 
-            if (_components != null) for (int i = 0; i < _components.Length; i++) if (_components[i] != null) _components[i].Destroy();
+            if (_components != null) for (int i = 0; i < _components.Length; i++) _components[i]?.Destroy();
             if (_transforms != null) {
                 for (int i = 0; i < _transforms.Length; i++) {
                     if (_transforms[i] != null && _transforms[i].gameObject != null) {
@@ -136,8 +135,9 @@ namespace Universum.World {
 
             _spinRotationSpeed = _rand.GetValueBetween(def.spinRotationSpeedBetween);
             float axialAngle = _rand.GetValueBetween(def.axialAngleBetween);
-            _axialRotation = Quaternion.Euler(0, axialAngle, 0);
-            _billboardCorrectionRotation = Quaternion.Euler(90, 90 + axialAngle, 0);
+            if (def.shape == null && def.icon != null) {
+                _axialRotation = Quaternion.Euler(0, 0, axialAngle);
+            } else _axialRotation = Quaternion.Euler(0, axialAngle, 0);
             float inclinationAngle = _rand.GetValueBetween(def.inclinationAngleBetween);
             if (inclinationAngle != 0) _inclinatioRotation = _rand.GetRotation() * Quaternion.Euler(_rand.GetValueBetween(def.inclinationAngleBetween), 0, 0);
 
@@ -172,28 +172,29 @@ namespace Universum.World {
         public virtual void UpdatePosition(int tick) {
             _positionChanged = true;
 
-            float time = speed * tick + _timeOffset;
-            float angularFrequencyTime = 6.28f / _period * time;
+            double time = speed * tick + _timeOffset;
+            double angularFrequencyTime = 6.28 / _period * time;
 
-            position.x = _orbitDirection * _orbitRadius * (float) Math.Cos(angularFrequencyTime);
-            position.z = _orbitDirection * _orbitRadius * Mathf.Sqrt(1 - _orbitEccentricity * _orbitEccentricity) * (float) Math.Sin(angularFrequencyTime);
+            position.x = (float) (_orbitDirection * _orbitRadius * Math.Cos(angularFrequencyTime));
+            position.z = (float) (_orbitDirection * _orbitRadius * Math.Sqrt(1 - _orbitEccentricity * _orbitEccentricity) * Math.Sin(angularFrequencyTime));
 
-            _localPosition = _inclinatioRotation * (position + GetTargetPosition());
+            _localPosition = (_inclinatioRotation * position) + GetTargetPosition();
         }
 
         public virtual void UpdateRotation(int tick) {
             _rotationChanged = true;
 
             if (_addBillboardRotation) {
-                _billboardRotation = Utils.billboardRotation(
-                    position,
-                    targetPosition: Game.MainLoop.instance.cameraPosition,
-                    _billboardCorrectionRotation
-                );
-
-                _rotation = _billboardRotation;
+                _billboardRotation = Utils.billboardRotation();
+                _rotation = _billboardRotation * _axialRotation;
             } else {
-                _spinRotation = Quaternion.Euler(0.5f * _spinRotationSpeed * tick * _orbitDirection * -1, _spinRotationSpeed * tick * _orbitDirection * -1, 0);
+                double num1 = _spinRotationSpeed * tick * _orbitDirection * -1 * (Math.PI / 180);
+                Vector3 euler = new Vector3(
+                    (float) (0.5 * num1),
+                    (float) num1,
+                    0.0f
+                );
+                _spinRotation = Quaternion.Internal_FromEulerRad(euler);
 
                 _rotation = _axialRotation * _spinRotation;
             }
@@ -232,8 +233,8 @@ namespace Universum.World {
             if (totalTransforms <= 0) return;
 
             for (int i = 0; i < totalTransforms; i++) {
-                if (updatePosition) _transforms[i].localPosition = _localPosition;
-                if (updateRotation) _transforms[i].localRotation = _rotation;
+                if (updatePosition) _transforms[i].position = _localPosition;
+                if (updateRotation) _transforms[i].rotation = _rotation;
                 if (updateScale) _transforms[i].localScale = scale;
             }
         }
@@ -312,16 +313,15 @@ namespace Universum.World {
 
         public virtual void UpdateOrbitRadius() {
             Vector3 scaledOrbitOffset = GetTargetScale() * _orbitPathOffsetPercentage;
-            _orbitRadius = scaledOrbitOffset.x + (float) ((_rand.GetFloat() - 0.5f) * (scaledOrbitOffset.x * _orbitSpread));
+            _orbitRadius = scaledOrbitOffset.x + ((_rand.GetFloat() - 0.5) * (scaledOrbitOffset.x * _orbitSpread));
         }
 
         public virtual void UpdateScale() {
             _scaleChanged = true;
 
-            Vector3 orbitAroundScale = GetTargetScale();
-            scale.x = orbitAroundScale.x * _scalePercentage;
-            scale.y = orbitAroundScale.y * _scalePercentage;
-            scale.z = orbitAroundScale.z * _scalePercentage;
+            scale = GetTargetScale() * _scalePercentage;
+
+            if (def.minSize != null && scale.x < (float) def.minSize) scale = new Vector3((float) def.minSize, (float) def.minSize, (float) def.minSize);
         }
 
         public virtual void UpdateSpeed() {
@@ -332,8 +332,8 @@ namespace Universum.World {
             return _target?.transformedPosition ?? Vector3.zero;
         }
 
-        public virtual float GetTargetSpeed() {
-            return _target?.speed ?? 0.8f;
+        public virtual double GetTargetSpeed() {
+            return _target?.speed ?? 0.8;
         }
 
         public virtual Vector3 GetTargetScale() {