+++ Changelog.md
@@ -1,1 +1,8 @@
 # Changelog
+- Improve generator compatibility when loading removed celestial object defs
+- Cache utilities which might be called through threads
+- Improve billboarding rotation
+- Support custom parameters for the object holder settle command
+- Fix celestial object targeting
+- Add minScale property to CelestialObject def
+- Fix precision issues when calculating position/rotation during late-game